requirements.txt
PyYAML==6.0
prometheus-client==0.14.1
requests==2.28.1